+++ src/cmd/split.rs
@@ -4,34 +4,40 @@ use std::num::NonZeroUsize;
 use std::path::Path;
 
 use crate::config::{Config, Delimiter};
-use crate::read::segment_csv_file;
+use crate::read::{segment_csv_file, SegmentationOptions};
 use crate::util::{self, FilenameTemplate};
 use crate::CliResult;
 
 static USAGE: &str = "
-Splits the given CSV data into chunks.
+Splits the given CSV data into smaller files having a fixed number of
+rows given to -s, --size.
 
-The files are written to the directory given with the name '{start}.csv',
-where {start} is the index of the first record of the chunk (starting at 0).
+Target file can also be split into a given number of -c/--chunks.
+
+Files will be written in current working directory by default or in any directory
+given to -O/--out-dir (that will be created for your if necessary).
 
 Usage:
-    xan split [options] <outdir> [<input>]
+    xan split [options] [<input>]
     xan split --help
 
 split options:
-    -s, --size <arg>       The number of records to write into each chunk.
-                           [default: 4096]
-    -c, --chunks <n>       Divide the file into approximately <n> chunks having
-                           roughly the same number of records. Target file must be
-                           seekable (e.g. this will not work with stdin nor gzipped
-                           files).
-    --segments             When used with -c/--chunks, output the byte offsets of
-                           found segments insteads.
-    --filename <filename>  A filename template to use when constructing
-                           the names of the output files.  The string '{}'
-                           will be replaced by a value based on the value
-                           of the field, but sanitized for shell safety.
-                           [default: {}.csv]
+    -O, --out-dir <dir>        Where to write the chunks. Defaults to current working
+                               directory.
+    -S, --size <arg>           The number of records to write into each chunk.
+                               [default: 4096]
+    -c, --chunks <n>           Divide the file into at most <n> chunks having
+                               roughly the same number of records. Target file must be
+                               seekable (e.g. this will not work with stdin nor gzipped
+                               files).
+    --segments                 When used with -c/--chunks, output the byte offsets of
+                               found segments insteads.
+    -f, --filename <filename>  A filename template to use when constructing
+                               the names of the output files. The string '{}'
+                               will be replaced either by the index in original file of
+                               first row emitted when using -S/--size or by the chunk
+                               index when using -c/--chunks.
+                               [default: {}.csv]
 
 Common options:
     -h, --help             Display this message
@@ -45,7 +51,7 @@ Common options:
 #[derive(Clone, Deserialize)]
 struct Args {
     arg_input: Option<String>,
-    arg_outdir: String,
+    flag_out_dir: Option<String>,
     flag_size: NonZeroUsize,
     flag_chunks: Option<NonZeroUsize>,
     flag_segments: bool,
@@ -61,16 +67,18 @@ pub fn run(argv: &[&str]) -> CliResult<()> {
         if args.flag_segments {
             args.segments()
         } else {
-            unimplemented!()
+            args.split_by_segments()
         }
     } else {
-        args.sequential_split()
+        args.split_by_size()
     }
 }
 
 impl Args {
-    fn sequential_split(&self) -> CliResult<()> {
-        fs::create_dir_all(&self.arg_outdir)?;
+    fn split_by_size(&self) -> CliResult<()> {
+        if let Some(out_dir) = &self.flag_out_dir {
+            fs::create_dir_all(out_dir)?;
+        }
 
         let rconfig = self.rconfig();
         let mut rdr = rconfig.reader()?;
@@ -87,39 +95,71 @@ impl Args {
             wtr.write_byte_record(&row)?;
             i += 1;
         }
-        wtr.flush()?;
+
+        Ok(wtr.flush()?)
+    }
+
+    fn split_by_segments(&self) -> CliResult<()> {
+        if let Some(out_dir) = &self.flag_out_dir {
+            fs::create_dir_all(out_dir)?;
+        }
+
+        let rconfig = self.rconfig();
+        let mut reader = rconfig.seekable_reader()?;
+
+        let segments = segment_csv_file(
+            &mut reader,
+            SegmentationOptions::chunks(self.flag_chunks.unwrap().get()),
+        )?
+        .ok_or("could not segment the file properly!")?;
+
+        let mut reader = rconfig.reader()?;
+        let headers = reader.byte_headers()?.clone();
+
+        let mut record = csv::ByteRecord::new();
+        let mut writer = self.new_writer(&headers, 0)?;
+        let mut chunk: usize = 0;
+
+        while reader.read_byte_record(&mut record)? {
+            if record.position().unwrap().byte() >= segments[chunk].1 {
+                writer.flush()?;
+                chunk += 1;
+                writer = self.new_writer(&headers, chunk)?;
+            }
+
+            writer.write_byte_record(&record)?;
+        }
+
         Ok(())
     }
 
     fn segments(&self) -> CliResult<()> {
-        let rconfig = self.rconfig().flexible(true);
-        let target = rconfig.io_reader_for_random_access()?;
-        let mut reader = rconfig.csv_reader_from_reader(target);
+        let rconfig = self.rconfig();
+        let mut reader = rconfig.seekable_reader()?;
 
-        let offsets_opt = segment_csv_file(&mut reader, self.flag_chunks.unwrap().get(), 128, 8)?;
+        let segments = segment_csv_file(
+            &mut reader,
+            SegmentationOptions::chunks(self.flag_chunks.unwrap().get()),
+        )?
+        .ok_or("could not segment the file properly!")?;
 
-        match offsets_opt {
-            None => Err("could not find segments!")?,
-            Some(offsets) => {
-                let mut wtr = Config::new(&None).writer()?;
-                let mut record = csv::ByteRecord::new();
+        let mut wtr = Config::new(&None).writer()?;
+        let mut record = csv::ByteRecord::new();
 
-                record.push_field(b"from");
-                record.push_field(b"to");
+        record.push_field(b"from");
+        record.push_field(b"to");
 
-                wtr.write_byte_record(&record)?;
+        wtr.write_byte_record(&record)?;
 
-                for (f, t) in offsets {
-                    record.clear();
-                    record.push_field(f.to_string().as_bytes());
-                    record.push_field(t.to_string().as_bytes());
+        for (f, t) in segments {
+            record.clear();
+            record.push_field(f.to_string().as_bytes());
+            record.push_field(t.to_string().as_bytes());
 
-                    wtr.write_byte_record(&record)?;
-                }
-
-                wtr.flush()?;
-            }
+            wtr.write_byte_record(&record)?;
         }
+
+        wtr.flush()?;
 
         Ok(())
     }
@@ -127,10 +167,13 @@ impl Args {
     fn new_writer(
         &self,
         headers: &csv::ByteRecord,
-        start: usize,
+        id: usize,
     ) -> CliResult<csv::Writer<Box<dyn io::Write + Send + 'static>>> {
-        let dir = Path::new(&self.arg_outdir);
-        let path = dir.join(self.flag_filename.filename(&format!("{}", start)));
+        let dir = match &self.flag_out_dir {
+            Some(out_dir) => Path::new(out_dir),
+            None => Path::new(""),
+        };
+        let path = dir.join(self.flag_filename.filename(&format!("{}", id)));
         let spath = Some(path.display().to_string());
         let mut wtr = Config::new(&spath).writer()?;
         if !self.rconfig().no_headers {